Update Main UI and refactoring sidenav UI
The commit introduces UI changes including expanding/collapsing side bar menu, new flex and font CSS classes, as well as some refactoring. It also moves the fade-in-out animation to the core/animations directory for better modularity.
