+++ src/app/app.component.scss
@@ -1,40 +1,40 @@
 @use '@angular/material' as mat;
 @use 'styles' as *;
 
-$toolbar-height: 64px;
-$additionalPadding: 16px;
-$listPaddingSize: 16px;
-$imageSize: 100px;
-$subTextSize: 19px;
-$dividerMargin: 9px;
-
-.mat-drawer-container,
-.mat-drawer-content {
+:host {
+  display: flex;
+  flex-direction: column;
   height: 100%;
   width: 100%;
+}
+
+.sidenav-container {
+  flex: 1;
+}
+
+.mat-drawer-container {
   overflow: hidden;
 }
 
-.mat-mdc-list {
-  height: calc(100% - #{$imageSize} - #{$subTextSize} - #{$dividerMargin} - #{$listPaddingSize});
-}
-
-.drawer-content {
-  height: calc(100% - (#{$toolbar-height} + #{$additionalPadding}));
-  overflow: auto;
+.mat-drawer-content {
   padding: 8px;
+  margin-left: 55px;
 }
 
-.mat-toolbar {
-  padding-right: 0;
+.content-container {
+  min-height: 100vh;
+  box-sizing: border-box;
+  padding: 16px 32px;
+
+  &.extended {
+    padding: 16px;
+  }
 }
 
 .window-control-buttons {
   display: flex;
   align-items: center;
   height: 100%;
-  padding-left: 24px;
-  gap: 16px;
 }
 
 .window-control-buttons .mat-icon {
@@ -42,7 +42,7 @@ $dividerMargin: 9px;
   justify-content: center;
   display: flex;
   align-items: center;
-  width: 40px;
+  width: 32px;
   cursor: pointer;
 
   &:hover {
@@ -67,8 +67,17 @@ $dividerMargin: 9px;
   }
 }
 
-.sidenav {
-  padding: 20px;
+mat-sidenav[style] {
+  visibility: visible !important;
+}
+
+.logo {
+  padding-right: 16px;
+}
+
+.mat-drawer {
+  display: flex !important;
+  transform: none !important;
 }
 
 .draggable-header {
@@ -79,4 +88,48 @@ $dividerMargin: 9px;
 
 .no-drag {
   -webkit-app-region: no-drag;
+}
+
+.issue-container {
+  display: flex;
+  flex-direction: column;
+}
+
+.side-navigation-actions-container {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  padding: 0;
+}
+
+.mdc-list-item__secondary-text {
+  display: flex;
+  font-size: smaller;
+  justify-content: center;
+}
+
+.version-container {
+  display: flex;
+  justify-content: center;
+  padding-bottom: 8px;
+  position: relative;
+
+  > span {
+    position: absolute;
+    right: 2px;
+    bottom: 0;
+  }
+}
+
+.mdc-list-item--with-leading-icon .mdc-list-item__start {
+  margin-left: 16px;
+  margin-right: 24px;
+}
+
+.mat-mdc-list-item {
+  height: 56px !important;
+
+  & mat-icon {
+    color: white;
+  }
 }
