+++ src/app/components/mod-search/mod-search.component.html
@@ -8,7 +8,7 @@
           <img alt="" class="example-option-img" [ngSrc]="mod.image ?? ''" width="50" height="50" />
           <div style="display: flex; flex-direction: column; padding-left: 8px; overflow: auto">
             <span class="text-ellipsis">{{ mod.name }}</span>
-            <span class="x-small">{{ mod.kind }}</span>
+            <span class="font-x-small">{{ mod.kind }}</span>
           </div>
           <span class="spacer"></span>
           @if (isActiveAkiInstanceAvailable()) {
